+++ Plugins/DynamicRefs/Source/NexusDynamicRefs/Public/NDynamicRefsGameplayTags.h
@@ -1,0 +1,15 @@
+﻿// Copyright dotBunny Inc. All Rights Reserved.
+// See the LICENSE file at the repository root for more information.
+
+#pragma once
+
+#include "NativeGameplayTags.h"
+
+#if WITH_EDITOR
+
+UE_DECLARE_GAMEPLAY_TAG_EXTERN(NDynamicRefs_EditorOnly_A);
+UE_DECLARE_GAMEPLAY_TAG_EXTERN(NDynamicRefs_EditorOnly_B);
+UE_DECLARE_GAMEPLAY_TAG_EXTERN(NDynamicRefs_EditorOnly_C);
+
+#endif // WITH_EDITOR
+