+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `NValidationMacros` to support providing easy to use validation for Blueprint-facing methods.
 - `NGuardian` now supports automatically setting its baseline after a defined period (default 15 seconds).
 - `NCell` side-car data file now has an action button, that when clicked in the Content Browser will select the source world/level. 
+- `FGameplayTag` support to `NDynamicRefs`.
 
 ### Changed